+++ lib/pin_screen.dart
@@ -1,0 +1,3 @@
+library custom_pin_screen;
+
+export 'src/src.dart';